requirements.txt
pyautogui
opencv-python
pillow
python-dotenv
requests
